refactor(runner): split runner.go into focused domain files
runner.go retains the struct, NewRunner, and echo.
runner_env.go owns environ().
runner_exec.go owns Run, RunWithStdin, RunInDir.
runner_output.go owns RunOutput, RunQuiet, RunQuietWithStdin.

Mirrors the pre-existing test file split (_env, _exec, _output).
No behavior or public API changes.
